Cleanup and move files
