+++ .github/workflows/build.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Run linter [flake8]
         run: |
           # stop the build if there are Python syntax errors or undefined names
-          flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+          flake8 .
       
       - name: Run unit tests [pytest]
         run: |
