+++ .vscode/launch.json
@@ -2,25 +2,10 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "Python: focus_stack",
+      "name": "Python: QtStackImages",
       "type": "python",
       "request": "launch",
-      "module": "focus_stack",
-      "args": ["-i", "images/", "-o", "/home/noah/Pictures/Aligned.jpg", "-g", "5", "-l", "5"],
-      "console": "integratedTerminal"
-    },
-    {
-      "name": "Python: StackImages.py",
-      "type": "python",
-      "request": "launch",
-      "program": "focus_stack/StackImages.py",
-      "console": "integratedTerminal"
-    },
-    {
-      "name": "Python: QtStackImages.py",
-      "type": "python",
-      "request": "launch",
-      "program": "QtUi/main.py",
+      "program": "focus_stack/main.py",
       "console": "integratedTerminal"
     }
   ]
