--- QtUi/__main__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Makes the focus_stack directory executable as a script."""
-
-from .main import main
-
-if __name__ == "__main__":
-    main()
