+++ pyproject.toml
@@ -1,6 +1,4 @@
+
 [build-system]
-requires = [
-    "setuptools>=42",
-    "wheel"
-]
+requires = ["setuptools>=46.0", "wheel"]
 build-backend = "setuptools.build_meta"
