+++ src/focus_stack/ParametersPopUp.py
