+++ src/focus_stack/QThreadWorkers.py
