+++ src/focus_stack/algorithm.py
