+++ src/focus_stack/icons.py
